+++ R/utils.R
@@ -286,6 +286,8 @@ splot.colormean = function(...){
 #'   as legend names. Otherwise, (and if the number of expressions is over 5 or the length of any
 #'   expression is over 50 characters) expressions are replaced with numbers corresponding to their
 #'   entered position.
+#' @param limit.outliers logical; if \code{TRUE} (default), times over an upper bound for the given
+#'   expression will be set to that upper bound, removing aberant extremes.
 #' @param options a list of options to pass on to splot.
 #' @examples
 #' # increase the number of runs for more stable estimates
@@ -339,46 +341,59 @@ splot.colormean = function(...){
 #' }
 #' @export
 
-splot.bench=function(...,runs=20,runsize=200,cleanup=FALSE,print.names=FALSE,options=list()){
-  e=sapply(as.character(substitute(list(...)))[-1],function(t)parse(text=t))
+splot.bench = function(..., runs = 20, runsize = 200, cleanup = FALSE, print.names = FALSE,
+  limit.outliers = TRUE, options = list()){
+  e = sapply(as.character(substitute(list(...)))[-1], function(t) parse(text = t))
   e = e[!duplicated(names(e))]
-  es=length(e)
+  es = length(e)
   if(!es) stop('no expressions found', call. = FALSE)
-  ne=names(e)
-  seconds=matrix(NA,runs,es,dimnames=list(c(),ne))
-  rs=seq_len(runsize)
-  p=cleanup || (runs*runsize>999)
-  tryCatch(for(t in e) eval(t),
-    error=function(e)stop('one of your expressions breaks:\n',e,call.=FALSE))
-  ost=proc.time()[3]
-  cat('benchmarking',es,'expression(s) in chunks of',runsize,'per run... ')
-  if(p) cat('\nrun 0 of',runs)
-  for(r in seq_len(runs)){
-    col=sample(ne)
-    seconds[r,col]=vapply(col,function(t){
-      if(cleanup) gc(FALSE)
-      st=proc.time()[3]
-      for(i in rs) eval(e[t])
-      proc.time()[3]-st
-    },0)
-    if(p) cat('\rrun',r,'of',runs)
+  ne = names(e)
+  seconds = matrix(NA, runs, es, dimnames = list(NULL, ne))
+  rs = seq_len(runsize)
+  tryCatch(
+    for(t in e) eval(t),
+    error = function(e) stop('one of your expressions breaks:\n', e, call. = FALSE)
+  )
+  ost = proc.time()[3]
+  cat('benchmarking', es, 'expression(s) in chunks of', runsize, 'per run... \nrun 0 of', runs)
+  fun = function(e){
+    eval(e)
+    NULL
   }
-  if(p) cat('\r')
-  cat('finished',runs,'runs in',round(proc.time()[3]-ost,2),'seconds       \n\n')
+  for(r in seq_len(runs)){
+    for(f in sample(seq_len(es))){
+      if(cleanup) gc(FALSE)
+      st = proc.time()[[3]]
+      for(i in rs) fun(e[[f]])
+      seconds[r, f] = proc.time()[[3]] - st
+    }
+    cat('\rrun', r, 'of', runs)
+  }
+  cat('\rfinished', runs, 'runs in', round(proc.time()[3] - ost, 2), 'seconds       \n\n')
   cat('expressions:\n\n')
-  icn=seq_len(es)
-  ne=gsub('\n','\n   ',ne,fixed=TRUE)
-  for(i in icn) cat(i,'. ',ne[i],'\n',sep='')
+  icn = seq_len(es)
+  ne = gsub('\n', '\n   ', ne, fixed = TRUE)
+  for(i in icn) cat(i, '. ', ne[i], '\n', sep = '')
   cat('\n')
-  print(round(matrix(c(colSums(seconds),colMeans(seconds)),2,byrow=TRUE,
-    dimnames=list(c('total time (seconds)','mean time per run'),icn)),4))
-  if(!print.names) if(!missing(print.names) || es>5 || any(nchar(names(e))>50))
-    colnames(seconds)=icn
-  if(es == 1 && runs == 1) return(NULL)
-  title=paste('timing of',runs,'runs of',runsize,'calls each')
+  res = rbind(colSums(seconds), colMeans(seconds))
+  res = rbind(res, if(min(res[1,], na.rm = TRUE) == 0) res[1,] + 1 else res[1,] / min(res[1,], na.rm = TRUE))
+  dimnames(res) = list(c('total time (seconds)', 'mean time per run', 'times the minimum'), icn)
+  print(round(res, 4))
+  if(!print.names) if(!missing(print.names) || es > 5 || any(nchar(names(e)) > 50))
+    colnames(seconds) = icn
+  if(limit.outliers) for(f in seq_len(es)){
+    qr = quantile(seconds[, f], c(.25, .75), TRUE)
+    qrc = qr[2] + (qr[2] - qr[1]) * 1.5
+    seconds[seconds[, f] > qrc, f] = qrc
+  }
+  if(es == 1 && runs == 1) return(list(plot = NULL, summary = res))
+  title = paste('timing of', runs, 'runs of', runsize, 'calls each')
   if(nrow(seconds) == 1){
     options$x = colnames(seconds)
     seconds = seconds[1,]
   }
-  splot(seconds,title=title,labels.filter=FALSE,labels.trim=FALSE,options=options)
+  invisible(list(
+    plot = splot(seconds, title = title, labels.filter = FALSE, labels.trim = FALSE, options = options),
+    summary = res
+  ))
 }